#correcao controllers - correcao da controller de avaliacao carrinhocompra categoria endereco e mensagem

+++ Pede_RocaAPP/Pede_RocaAPP.API/Controllers/CategoriaController.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using Pede_RocaAPP.Application.DTOs;
 using Pede_RocaAPP.Application.Interface;
+using Pede_RocaAPP.Application.Services;
+using Pede_RocaAPP.Domain.Entities;
 
 namespace Pede_RocaAPP.API.Controllers
 {
@@ -16,44 +18,74 @@ namespace Pede_RocaAPP.API.Controllers
         }
 
         [HttpPost(Name = "AdicionarCategoria")]
-        public async Task<ActionResult> Post([FromBody] CategoriaDTO categoriaDTO)
+        public async Task<ActionResult> Post([FromBody] CategoriaCreateDTO categoriaDTO)
         {
             if (categoriaDTO == null)
             {
-                return BadRequest("Erro de dado inválido");
+                return BadRequest("Erro de dados inválidos. Verifique o payload de envio e tente novamente!");
             }
-            await _categoriaService.AdicionarAsync(categoriaDTO);
 
-            return CreatedAtRoute("GetCategoria", new { id = categoriaDTO.Id }, categoriaDTO);
+            var categoriaId = await _categoriaService.AdicionarAsync(categoriaDTO);
+
+            return CreatedAtRoute("GetCategoria", new { id = categoriaId }, new
+            {
+                id = categoriaId,
+                message = "Categoria criada com sucesso"
+            });
         }
 
         [HttpPut("{id}", Name = "AtualizarCategoria")]
-        public async Task<ActionResult> Put(Guid id, [FromBody] CategoriaDTO categoriaDTO)
+        public async Task<ActionResult> Put(Guid id, [FromBody] CategoriaCreateDTO categoriaDTO)
         {
-            if (id != categoriaDTO.Id)
-            {
-                return BadRequest("Id não válido");
-            }
             if (categoriaDTO == null)
             {
-                return BadRequest("Erro de dado inválido");
+                return BadRequest("Erro de dados inválidos. Verifique o payload de envio e tente novamente!");
             }
 
-            await _categoriaService.AtualizarAsync(id, categoriaDTO);
-            return Ok(categoriaDTO);
+            var categoriaEncontrada = await _categoriaService.GetByIdUpdateAsync(id);
+
+            if (categoriaEncontrada == null)
+            {
+                return NotFound($"Categoria com ID {id} não encontrada. Verifique o ID e tente novamente!");
+            }
+
+            if(!string.IsNullOrEmpty(categoriaDTO.Nome))
+            {
+                categoriaEncontrada.Nome = categoriaDTO.Nome;
+            }
+
+            await _categoriaService.AtualizarAsync(id, categoriaEncontrada);
+            return Ok(new
+            {
+                mensagem = $"Categoria com o id {id} foi atualizada com sucesso"
+            });
         }
 
         [HttpDelete("{id}", Name = "DeleteCategoria")]
+        [ProducesResponseType(StatusCodes.Status200OK, Type = typeof(MensagemResponse))]
         public async Task<ActionResult<CategoriaDTO>> DeleteAsync(Guid id)
         {
             var categoriaDto = await _categoriaService.GetByIdAsync(id);
+
             if (categoriaDto == null)
             {
                 return NotFound("Categoria não encontrada");
             }
+
+            // Verifica se há produtos associados à categoria
+            var produtosAssociados = await _categoriaService.GetByCategoriaIdAsync(id);
+
+            if (produtosAssociados != null && produtosAssociados.Any())
+            {
+                return BadRequest("Não é possível excluir a categoria, pois existem produtos associados a ela.");
+            }
+
             await _categoriaService.DeleteAsync(id);
 
-            return Ok(categoriaDto);
+            return Ok(new
+            {
+                message = "Categoria removida com sucesso"
+            });
         }
 
         [HttpGet("{id}", Name = "GetCategoria")]
